+++ src/main/java/com/carolin_violet/travel_system/controller/Front/ItemOrderController.java
@@ -9,13 +9,13 @@ import com.carolin_violet.travel_system.bean.vo.OrderVo2;
 import com.carolin_violet.travel_system.service.OrderFormService;
 import com.carolin_violet.travel_system.service.ScenicSpotService;
 import com.carolin_violet.travel_system.service.TouristRouteService;
+import com.carolin_violet.travel_system.utils.JwtUtils;
 import com.carolin_violet.travel_system.utils.R;
-import com.sun.org.apache.xpath.internal.operations.Or;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.annotation.Order;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -44,7 +44,13 @@ public class ItemOrderController {
 
     // 游客查询没支付的订单
     @GetMapping("getUnPaidOrder/{touristId}")
-    public R getUnPaidOrder(@PathVariable String touristId) {
+    public R getUnPaidOrder(@PathVariable String touristId, HttpServletRequest httpServletRequest) {
+
+        // 身份验证
+        if (!JwtUtils.getMemberIdByJwtToken(httpServletRequest).equals(touristId)) {
+            return R.error().message("身份错误！");
+        }
+
         // 游客查询到自己所有的订单
         QueryWrapper<OrderForm> wrapper = new QueryWrapper<>();
         wrapper.eq("tourist_id", touristId).eq("is_paid", 0);
@@ -74,7 +80,14 @@ public class ItemOrderController {
 
     // 游客根据订单时间倒序分页查询所有历史订单
     @GetMapping("order/{touristId}/{cur}/{limit}")
-    public R getPageOrder(@PathVariable String touristId, @PathVariable long cur, @PathVariable long limit) {
+    public R getPageOrder(@PathVariable String touristId, @PathVariable long cur, @PathVariable long limit, HttpServletRequest httpServletRequest) {
+
+        // 身份验证
+        if (!JwtUtils.getMemberIdByJwtToken(httpServletRequest).equals(touristId)) {
+            return R.error().message("身份错误！");
+        }
+
+
         Page<OrderForm> orderFormPage = new Page<>(cur, limit);
         QueryWrapper<OrderForm> wrapper = new QueryWrapper<>();
         wrapper.orderByDesc("create_time").eq("tourist_id", touristId);
@@ -82,9 +95,6 @@ public class ItemOrderController {
 
         long total = orderFormPage.getTotal();
         List<OrderForm> records = orderFormPage.getRecords();
-        System.out.println("6666666666666666666666666666666666666666666");
-        System.out.println(touristId);
-        System.out.println(records.toString());
 
         // 每一个订单再查询对应的名称(景点、线路拼团)
         List<OrderVo2> orderList = new ArrayList<>();
@@ -109,7 +119,13 @@ public class ItemOrderController {
 
     // 游客新增订单，可以是拼单也可以是单买门票
     @PostMapping("addOrder")
-    public R addOrder(@RequestBody OrderForm orderForm) {
+    public R addOrder(@RequestBody OrderForm orderForm, HttpServletRequest httpServletRequest) {
+
+        // 身份验证
+        if (!JwtUtils.getMemberIdByJwtToken(httpServletRequest).equals(orderForm.getTouristId())) {
+            return R.error().message("身份错误！");
+        }
+
         boolean save = orderFormService.save(orderForm);
         if (save) {
             return R.ok();
@@ -120,7 +136,13 @@ public class ItemOrderController {
 
     // 修改订单信息,支付后修改订单信息为支付成功
     @PutMapping("updateOrder")
-     public R updateOrder(@RequestBody OrderForm orderForm) {
+     public R updateOrder(@RequestBody OrderForm orderForm, HttpServletRequest httpServletRequest) {
+
+        // 身份验证
+        if (!JwtUtils.getMemberIdByJwtToken(httpServletRequest).equals(orderForm.getTouristId())) {
+            return R.error().message("身份错误！");
+        }
+
         boolean flag = orderFormService.updateById(orderForm);
         if (flag) {
             return R.ok();
@@ -131,7 +153,12 @@ public class ItemOrderController {
 
     // 游客删除自己的订单
     @DeleteMapping("delOrder/{id}")
-    public R delOrder(@PathVariable String id) {
+    public R delOrder(@PathVariable String id, HttpServletRequest httpServletRequest) {
+
+        if (!JwtUtils.getMemberIdByJwtToken(httpServletRequest).equals(orderFormService.getById(id).getTouristId())) {
+            return R.error().message("身份错误！");
+        }
+
         boolean flag = orderFormService.removeById(id);
         if (flag) {
             return R.ok();
